Move code to separated file
